기능추가: 게시글 삭제 구현

+++ src/main/java/com/ksy/noticeboard/repository/BoardRepository.java
@@ -13,7 +13,7 @@ public interface BoardRepository {
     List<BoardVO> findAllByPagination(Pagination pagination);
     void create(Board board);
     void update(Board board);
-    void delete(Long id);
+    void delete(int id);
     int countAll();
     BoardVO findById(int id);
 }
